refactor(dashboard): basic dashboard table component integration is done
- File .env was added to gitignore.

- New constants files were created.
